Split file modal functions.
