+++ frontend/tsconfig.json
@@ -12,6 +12,7 @@
   },
   "files": [
     "src/helper.ts",
+    "src/modal.ts",
     "src/server_status.ts",
     "src/users_admin.ts",
     "src/models/base.ts",
